feat: display config state for create record
